Add files via upload
This is the core JS. Note: This is build is missing the icon and any card assets tested with this. I will not be uploading card assets here as that could conflict with distribution rights of the TTRPG's or other systems involved. Please create or find assets from another source.

Another recommendation is keeping track of stat blocks as a DM/GM/Story Teller in here as you can quickly tab through them.